problemas

+++ src/main/java/com/desafio/codenation/repositories/SistemaRepositorie.java
@@ -4,5 +4,5 @@ import com.desafio.codenation.domain.origem.Sistema;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SistemaRepositorie extends BaseRepository<Sistema, Integer> {
+public interface SistemaRepositorie extends BaseRepository<Sistema, Long> {
 }
